+++ src/status.rs
@@ -6,20 +6,31 @@ use std::fmt;
 /// This contains the payload that Home Assistant uses to read the values.
 #[derive(Serialize, Debug, Default)]
 pub struct StatusMessage {
+    pub available: &'static str,
+
     /// CPU usage in %
-    pub cpu_usage: f32,
+    pub cpu_usage: Option<f32>,
 
     /// CPU temperature in °C
     pub cpu_temp: Option<f32>,
 
     /// Memory usage in %
-    pub memory_usage: f32,
+    pub memory_usage: Option<f32>,
 
     /// Net TX rate in KiB/s
     pub net_tx: Option<f64>,
 
     /// Net RX rate in KiB/s
     pub net_rx: Option<f64>,
+}
+
+impl StatusMessage {
+    pub fn off() -> StatusMessage {
+        StatusMessage {
+            available: "OFF",
+            ..Default::default()
+        }
+    }
 }
 
 impl fmt::Display for StatusMessage {